Week6
Important notes: [..arrayinitial, prop added to array], note that arrays are immutable therefore, in order to render them, you are adding to an array, but you are in a way, rendering the added item/prop alongside the original array.  In the new-item.js you had to create the function that calls onAddItem which is invoked on the page.js. This is so you can take the prop with the item object from the page and add it to the form.  Item-list is responsible for the states, From limes36,46 and the mapping from 109 and 125. What is happening is you are calling the original array, but as well you are updating the state with the new prop that was added such that when sorted, you are also including the new prop.  Page.js is 100% responsible for handling the new prop. It takes the new item and adds it to the very end of the Array. But it also initializes the state.