Week 8: firebase auth
Worked on setting up Firebase authentication; setting up my account connecting github. Using Context provider to use the context/access and pass it. auth-context is used to handle the state of authentication and the user.  Firebase is used to get access and compare. page.js is the front-end styling, has some additional security practice that prohibits the user seeing the shopping list.  If user exists then shopping List is a valid option in week/8/shopping-list.